Reinstate cc7/js/Utils as CC7Utils
